a bit better texture rounding

+++ src/render.c
@@ -6,7 +6,7 @@
 /*   By: malaakso <malaakso@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/13 18:19:30 by malaakso          #+#    #+#             */
-/*   Updated: 2023/11/02 16:33:36 by malaakso         ###   ########.fr       */
+/*   Updated: 2023/11/02 16:38:04 by malaakso         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -91,7 +91,7 @@ void	draw_texture(t_data *d, t_ray *ray, int x)
 		ray->tex_pos.x = ray->texture->width - ray->tex_pos.x - 1;
 	step = 1.0 * ray->texture->height / ray->line_height;
 	tex_start_pos = (ray->draw_start
-			- WINDOW_HEIGHT / 2 + ray->line_height / 2) * step;
+			- (double)WINDOW_HEIGHT / 2 + (double)ray->line_height / 2) * step;
 	y = ray->draw_start;
 	while (y <= ray->draw_end)
 	{
